pyproject.toml
[tool.poetry]
name = "lsu-pilot"
version = "0.1.0"
description = ""
authors = ["Daniel Williams <dannyhyunsoowilliams@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
python-telegram-bot = "^21.1.1"
openai = "^1.25.1"
python-dotenv = "^1.0.1"
langchain = "^0.1.17"
tiktoken = "^0.6.0"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
